Replace inactive echo service in websocket.rst
The recommended websocket echo service (ws://echo.websocket.org) is no longer available, so the minimal example given will not work. I've proposed the first available one I found, but it is not ideal - while it does echo messages sent to it, it will also send other messages to connected clients. This is nevertheless an improvement on a non-functioning link.